requirements.txt
# ipwndfu - iOS exploit tool dependencies
# Python 3.9+ required

# USB communication library
pyusb>=1.2.1

# System utilities
setuptools>=45.0.0
wheel>=0.36.0

# Optional: For enhanced functionality
cryptography>=3.4.8

# Development dependencies (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.8.0
